functions/requirements.txt
firebase_functions~=0.5.0
google-cloud-pubsub>=2.0.0
google-cloud-logging>=3.0.0